requirements.txt
# Crypto Trading Platform — dependencies
# Python 3.9+ recommended

PyQt6>=6.4.0
python-binance>=1.0.19
requests>=2.28.0
websocket-client>=1.5.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.1.0
